moved blockscanner to the separate file
